Speed up quote addition using root key
Stores next quote id in a root-level key in order to speed up the addition of quotes. 

Even if a user's json is freshly exported from streamerbot and does not contain this id, then the id will be created on the next quote addition, meaning their quotes json will automatically update and be quicker upon subsequent adds.

This is a speedup to ensure we keep light arch here and just use basic file i/o. If a need for a db or other type of persistence arises...then this project has grown much bigger than anticipated.